+++ lib/src/features/cipher/presentation/cipher_mode_state.g.dart
@@ -6,7 +6,7 @@ part of 'cipher_mode_state.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$cipherModeStateHash() => r'aa76dd30f413e5e9ef9de5410c3cee0f3bed604d';
+String _$cipherModeStateHash() => r'19918dd1c56052e4746085aa2864db801ac98a2c';
 
 /// See also [CipherModeState].
 @ProviderFor(CipherModeState)