+++ lib/src/features/cipher/presentation/components/algorithm_list_tile.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter/material.dart';
 
+import '../../../../common_widgets/custom_button.dart';
 import '../../../../constants/_constants.dart';
 import '../../../../core/_core.dart';
 import '../../../../util/context_shortcut.dart';
@@ -17,7 +18,7 @@ class AlgorithmListTile extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final cipherMode = ref.watch(cipherModeStateProvider);
-// TODO default set algorithm
+
     return Column(
       children: [
         GAP_H2,
@@ -29,11 +30,12 @@ class AlgorithmListTile extends ConsumerWidget {
                   fontWeight: FontWeight.bold,
                 ),
           ),
+          subtitle: Text(currentAlgorithm.detail),
           selected: cipherMode.algorithm == currentAlgorithm,
           selectedTileColor:
               (cipherMode.algorithm == currentAlgorithm) ? PRIMARY_COLOR_D0.withAlpha(50) : null,
-          subtitle: Text(currentAlgorithm.detail),
           leading: Icon(currentAlgType.icon),
+          trailing: AlgorithmListTileTrailing(currentAlgorithm: currentAlgorithm),
           tileColor: kColor(context).surfaceDim,
           dense: true,
           onTap: () {
@@ -56,6 +58,57 @@ class AlgorithmListTile extends ConsumerWidget {
           },
         ),
       ],
+    );
+  }
+}
+
+class AlgorithmListTileTrailing extends ConsumerWidget {
+  const AlgorithmListTileTrailing({super.key, required this.currentAlgorithm});
+  final CipherAlgorithm currentAlgorithm;
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return FutureBuilder(
+      future: ref.read(cipherModeStateProvider.notifier).getDefault(),
+      builder: (context, snapshot) {
+        if (snapshot.connectionState == ConnectionState.waiting) {
+          return const Padding(
+            padding: EdgeInsets.all(5),
+            child: Icon(Icons.star_border_rounded, size: 20),
+          );
+        }
+
+        final defaultCipherMode = snapshot.data;
+
+        return CustomButton(
+          msg: "Set to default",
+          buttonColor: Colors.transparent,
+          padding: EdgeInsets.all(5),
+          borderRadius: BorderRadius.circular(60),
+          onTap: () async {
+            if (!currentAlgorithm.supported) {
+              ScaffoldMessenger.of(context).clearSnackBars();
+              ScaffoldMessenger.of(context).showSnackBar(
+                const SnackBar(
+                  content: Text("Sorry, not supported yet! :)"),
+                  dismissDirection: DismissDirection.horizontal,
+                ),
+              );
+              return;
+            }
+
+            if (defaultCipherMode == currentAlgorithm) return;
+
+            // Set to default
+            await ref.read(cipherModeStateProvider.notifier).setDefault(currentAlgorithm);
+            ref.invalidate(cipherModeStateProvider);
+          },
+          child: Icon(
+            (defaultCipherMode == currentAlgorithm) ? Icons.star_rounded : Icons.star_border_rounded,
+            size: 20,
+          ),
+        );
+      },
     );
   }
 }